+++ marshw.h
@@ -124,7 +124,7 @@ void Mars_ClearNTA(void);
 void Mars_MDPutString(char *str);
 
 void Mars_DebugStart(void);
-void Mars_DebugQueue(int id, int val);
+void Mars_DebugQueue(int id, short val);
 void Mars_DebugEnd(void);
 
 void Mars_SetBankPage(int bank, int page) MARS_ATTR_DATA_CACHE_ALIGN;
